+++ sdk/monitor/query/azlogs/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Release History
 
-## 1.0.0 (2024-04-10)
+## 1.0.0 (2024-04-15)
 
 ### Breaking Changes
 * Removed pointers from slices
+* Renamed `LogsColumnType` to `ColumnType`
+* Renamed `LogsQueryOptions` to `QueryOptions`
+* Renamed `ServiceNameLogs` to `ServiceName`
 
 ## 0.1.0 (2024-03-07)